+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `STEP` endpoint to manually step the simulation forwards when in agent mode
 - `SIMMODE` endpoint to allow switching between modes at runtime
 - `SEED` endpoint to allow setting BlueSky's random seed
+- `LOADLOG` endpoint to reset the sim to a given point from a logfile
 - Added basic aircraft separation metrics
 - Added `GET` option to the `ALT` endpoint, to return the 3 versions of the flight levels
 
@@ -21,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Bump required BlueSky version to `1.3.0-dev`
 - When in `agent` mode, aircraft data is only logged after each `STEP` command
+- Point BlueSky submodule at `tags/turing-1.2.1`
 
 ### Removed